Step 8, Added users and groups with permissions
Added groups that are for people that are loaning books out and another group for librarians. These groups have different permission and show different information based on the user group. then added pages that show the books a user has loaned out and a page for librarians that shows all books that are loaned out